requirements.txt
python-telegram-bot==21.*
Pillow==10.*
requests==2.*
